requirements.txt
web3
mnemonic
eth-account
python-dotenv